Rename : 등록된 카드 이미지 관련 패키지 이동
1. 등록된 카드 이미지 - ui.card.list.component.card
