+++ app/src/main/java/nextstep/payments/ui/card/list/component/card/CardOwnerName.kt
@@ -1,13 +1,10 @@
-package nextstep.payments.ui.card.list.component
+package nextstep.payments.ui.card.list.component.card
 
-import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.unit.dp
-import nextstep.payments.data.Card
 
 @Composable
 fun CardOwnerName(ownerName: String, modifier: Modifier = Modifier) {
